+++ src/Graphics/Graphics.cs
@@ -76,7 +76,7 @@ public static class Graphics
         if (initialized)
             return;
 
-        if (SDL.SDL_Init(SDL.SDL_INIT_VIDEO) < 0)
+        if (SDL.SDL_Init(SDL.SDL_INIT_VIDEO | SDL.SDL_INIT_AUDIO) < 0)
         {
             throw new Exception($"SDL could not initialize! SDL Error: {SDL.SDL_GetErrorString()}");
         }